Splits ConstraintSolverConf out to its own files.
