+++ PlayRho/Dynamics/Contacts/ContactSolver.cpp
@@ -18,6 +18,8 @@
  */
 
 #include <PlayRho/Dynamics/Contacts/ContactSolver.hpp>
+
+#include <PlayRho/Dynamics/Contacts/ConstraintSolverConf.hpp>
 #include <PlayRho/Collision/Collision.hpp>
 #include <PlayRho/Collision/WorldManifold.hpp>
 #include <PlayRho/Dynamics/Contacts/BodyConstraint.hpp>
@@ -470,26 +472,6 @@ inline Momentum SolveNormalConstraint(VelocityConstraint& vc,
 }; // anonymous namespace
     
 } // namespace d2
-
-ConstraintSolverConf GetRegConstraintSolverConf(const StepConf& conf) noexcept
-{
-    return ConstraintSolverConf{}
-        .UseResolutionRate(conf.regResolutionRate)
-        .UseLinearSlop(conf.linearSlop)
-        .UseAngularSlop(conf.angularSlop)
-        .UseMaxLinearCorrection(conf.maxLinearCorrection)
-        .UseMaxAngularCorrection(conf.maxAngularCorrection);
-}
-
-ConstraintSolverConf GetToiConstraintSolverConf(const StepConf& conf) noexcept
-{
-    return ConstraintSolverConf{}
-        .UseResolutionRate(conf.toiResolutionRate)
-        .UseLinearSlop(conf.linearSlop)
-        .UseAngularSlop(conf.angularSlop)
-        .UseMaxLinearCorrection(conf.maxLinearCorrection)
-        .UseMaxAngularCorrection(conf.maxAngularCorrection);
-}
 
 namespace GaussSeidel {
