+++ PlayRho/Dynamics/Joints/RopeJointConf.cpp
@@ -25,7 +25,7 @@
 #include <PlayRho/Dynamics/Joints/Joint.hpp>
 #include <PlayRho/Dynamics/StepConf.hpp>
 #include <PlayRho/Dynamics/Contacts/BodyConstraint.hpp>
-#include <PlayRho/Dynamics/Contacts/ContactSolver.hpp> // for ConstraintSolverConf
+#include <PlayRho/Dynamics/Contacts/ConstraintSolverConf.hpp>
 
 namespace playrho {
 namespace d2 {
